Remove some unused functions (collect_operands, is_well_formed_func)
Move some syntax-related definitions into syntax.rs
