+++ Cargo.toml
@@ -17,6 +17,5 @@ js-sys = "0.3.81"
 str_indices = "0.4.4"
 wasm-bindgen = "0.2.104"
 web-sys = { version = "0.3.81", features = ["Text", "Element", "HtmlDivElement", "Window", "Document", "console", "HtmlBodyElement", "NodeList", "HtmlBrElement", "HtmlSpanElement", "HtmlParagraphElement", "HtmlElement", "InputEvent", "Range", "Selection", "DataTransfer", "KeyboardEvent", "ScrollIntoViewOptions", "ScrollBehavior", "ScrollLogicalPosition", "SvgElement", "SvgLineElement", "SvgDefsElement", "SvgMarkerElement", "SvggElement", "SvgCircleElement", "SvgUseElement", "SvgPathElement", "Performance"] }
-self_cell = "1.2.0"
 wasm-bindgen-futures = "0.4.54"
 wasm-encoder = { version = "0.239.0", features = ["wasmparser"] }
