+++ src/utils.rs
@@ -1,9 +1,4 @@
-use crate::{
-    line::{Activity, LineInfo},
-    syntax::{InstrKind, SyntheticWasm},
-};
 use anyhow::{Result, anyhow};
-use std::ops::{Deref, RangeInclusive};
 use wasm_encoder::{
     CodeSection, ExportSection, FunctionSection, GlobalSection, Instruction as EncoderInstruction,
     TypeSection,
@@ -306,12 +301,12 @@ pub struct TypesTable {
     _table: Vec<CodillonType>,
 }
 
-pub fn str_to_binary(s: String) -> Result<Vec<u8>> {
-    let txt = format!("module {s}");
-    let binding = ParseBuffer::new(&txt)?;
-    let mut module = parser::parse::<Module>(&binding)?;
+pub fn str_to_binary(mut txt: String) -> Result<Vec<u8>> {
+    txt.insert_str(0, "module ");
+    let wasm_bin = parser::parse::<Module>(&ParseBuffer::new(&txt)?)?.encode()?;
+    parse_binary_wasm(wasmparser::Parser::new(0), &wasm_bin)?; // make sure binary is well-formed
 
-    Ok(module.encode()?)
+    Ok(wasm_bin)
 }
 
 // Find the line comment separator in these string slices.
@@ -323,131 +318,6 @@ pub fn find_comment(s1: &str, s2: &str) -> Option<usize> {
     } else {
         s2.find(";;").map(|idx| s1.len() + idx)
     }
-}
-
-/// Decides if a given string is a well-formed text-format Wasm function
-///
-/// Uses wast ParseBuffer to convert string into buffer and wast parser to parse buffer as Module
-/// Encodes Module to binary Wasm and wasmparser parses binary Wasm
-///
-/// # Parameters
-/// lines: A string slice representing a Wasm function
-///
-/// # Returns
-/// true: if the function is syntactically well-formed; false otherwise
-///
-/// # Assumptions
-/// Each instruction is plain
-pub fn is_well_formed_func(lines: &str) -> bool {
-    let parse_text = || {
-        let text = format!("module (func {lines})");
-        let binary = parser::parse::<Module>(&ParseBuffer::new(&text)?)?.encode()?;
-        parse_binary_wasm(wasmparser::Parser::new(0), &binary)
-    };
-    parse_text().is_ok()
-}
-
-/// Return value of collect_operands - represents params and results of each instruction
-pub type InstrOps = Vec<(Vec<(ValType, usize)>, Vec<ValType>)>;
-
-/// Returns input and output types for each instruction in a binary Wasm module
-///
-/// # Parameters
-/// wasm_bin: a binary Wasm module
-/// ops: a vector of Operators and their idx in the editor
-///
-/// # Returns
-/// Err if function is not valid, else:
-/// A vector of tuples where:
-/// the first element is a (ValType, usize) tuple representing an operand this instruction pops and the idx in the editor where this operand was pushed
-/// the second element is a ValType representing an operand this instruction pushes
-// TODO: use Operators in OkModule instead of re-parsing the binary, and annotate each Operator in OkModule with its type instead of returning a vector
-pub fn collect_operands<'a>(wasm_bin: &[u8], ops: &Vec<(Operator<'a>, usize)>) -> Result<InstrOps> {
-    let mut validator = Validator::new();
-    let parser = wasmparser::Parser::new(0);
-    let mut result = Vec::new();
-    let dummy_offset = 1; // no need to track offsets, but validator has safety checks against 0
-
-    for payload in parser.parse_all(wasm_bin) {
-        if let ValidPayload::Func(func, _) = validator.payload(&payload?)? {
-            let mut func_validator: wasmparser::FuncValidator<wasmparser::ValidatorResources> =
-                func.into_validator(wasmparser::FuncValidatorAllocations::default());
-            let mut idx_stack: Vec<usize> = Vec::new(); // simulated operand stack to track idx where each operand is pushed
-            for (op, idx) in ops {
-                let (pop_count, push_count) = op
-                    .operator_arity(&func_validator.visitor(dummy_offset))
-                    .ok_or(anyhow!("could not determine operator arity"))?;
-                let prev_height = func_validator.operand_stack_height();
-                if pop_count > prev_height {
-                    return Err(anyhow!("expected to pop operand, but empty stack"));
-                }
-                let mut inputs = (prev_height - pop_count..prev_height)
-                    .filter_map(|i| {
-                        let valtype = func_validator.get_operand_type(i as usize).flatten();
-                        let idx = idx_stack.pop();
-                        match (valtype, idx) {
-                            (Some(val), Some(idx)) => Some((val, idx)),
-                            _ => None,
-                        }
-                    })
-                    .collect::<Vec<_>>();
-
-                //reverse inputs so that operands pushed more recently are towards the left (for UI)
-                inputs.reverse();
-
-                func_validator.op(dummy_offset, op)?;
-
-                let new_height = func_validator.operand_stack_height();
-                if push_count > new_height {
-                    return Err(anyhow!(
-                        "expected operator to push operand that is not on the stack"
-                    ));
-                }
-                let outputs = (new_height - push_count..new_height)
-                    .filter_map(|i| func_validator.get_operand_type(i as usize).flatten())
-                    .collect::<Vec<_>>();
-                for _ in &outputs {
-                    idx_stack.push(*idx);
-                }
-
-                result.push((inputs, outputs));
-            }
-        }
-    }
-
-    Ok(result)
-}
-
-/// Represents a frame entry (like "block end" pair, etc.), with range recorded.
-/// The range is inclusive, containing both start instr number and end instr number.
-/// The start number begins at 0.
-pub type Frame = RangeInclusive<usize>;
-
-pub trait LineInfos {
-    fn is_empty(&self) -> bool;
-    fn len(&self) -> usize;
-    fn info(&self, index: usize) -> impl Deref<Target = LineInfo>;
-}
-
-pub trait LineInfosMut: LineInfos {
-    fn set_active_status(&mut self, index: usize, new_val: Activity);
-    fn set_synthetic_before(&mut self, index: usize, synth: SyntheticWasm);
-    fn push(&mut self);
-}
-
-#[derive(PartialEq, Clone, Debug)]
-pub struct FrameInfo {
-    pub indent: usize,
-    pub start: usize,
-    pub end: usize,
-    pub unclosed: bool,
-    pub kind: InstrKind,
-}
-
-pub trait FrameInfosMut: LineInfos {
-    fn set_indent(&mut self, index: usize, num: usize);
-    fn set_frame_count(&mut self, count: usize);
-    fn set_frame_info(&mut self, num: usize, frame: FrameInfo);
 }
 
 pub trait FmtError {
@@ -467,7 +337,11 @@ pub(crate) mod tests {
     use wasmparser::BlockType;
 
     #[test]
-    fn test_is_well_formed_func() {
+    fn test_str_to_binary_for_one_function() {
+        fn is_well_formed_func(s: &str) -> bool {
+            str_to_binary(format!("(func {s})")).is_ok()
+        }
+
         //well-formed function
         assert!(is_well_formed_func("block\nend\n"));
         assert!(is_well_formed_func("i32.const 1\ni32.const 2\ni32.add"));
@@ -494,7 +368,7 @@ pub(crate) mod tests {
     }
 
     #[test]
-    fn test_collect_operands() -> Result<()> {
+    fn test_types_table() -> Result<()> {
         //block instruction with params and results
         let output = TypesTable {
             _table: vec![
